feat: add order validation before accepting order

+++ bar-app/src/main/java/by/pub/bar/app/order/service/OrderServiceImpl.java
@@ -2,25 +2,32 @@ package by.pub.bar.app.order.service;
 
 import by.pub.bar.app.event.entity.NewOrderSavedEvent;
 import by.pub.bar.app.event.publisher.BarEventPublisher;
+import by.pub.bar.app.ingredient.service.IngredientService;
 import by.pub.bar.app.order.entity.Order;
 import by.pub.bar.app.order.repository.OrderRepository;
 import by.pub.bar.app.order.utils.OrderDBProcessor;
+import by.pub.bar.app.product.service.ProductService;
 import by.pub.bar.app.utils.Status;
 import by.pub.bar.app.websocket.server.message_sender.OrderWebSocketMessageSender;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
 public class OrderServiceImpl implements OrderService {
     private final OrderRepository orderRepository;
+    private final ProductService productService;
+    private final IngredientService ingredientService;
     private final OrderDBProcessor processor;
     private final OrderWebSocketMessageSender orderWebSocketMessageSender;
     private final BarEventPublisher publisher;
 
-    public OrderServiceImpl(OrderRepository orderRepository, OrderDBProcessor processor, OrderWebSocketMessageSender orderWebSocketMessageSender, BarEventPublisher publisher) {
+    public OrderServiceImpl(OrderRepository orderRepository, ProductService productService, IngredientService ingredientService, OrderDBProcessor processor, OrderWebSocketMessageSender orderWebSocketMessageSender, BarEventPublisher publisher) {
         this.orderRepository = orderRepository;
+        this.productService = productService;
+        this.ingredientService = ingredientService;
         this.processor = processor;
         this.orderWebSocketMessageSender = orderWebSocketMessageSender;
         this.publisher = publisher;
@@ -42,6 +49,9 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     public Order saveOrder(Order order) {
+        order.setProducts(order.getProducts().stream()
+                .map(product -> productService.findByName(Objects.requireNonNull(product.getName())))
+                .collect(Collectors.toList()));
         final Order savedOrder = processor.fromDB(orderRepository.save(processor.toDB(processor.preprocessTotalPrice(order))));
         publisher.publishEvent(new NewOrderSavedEvent(savedOrder));
         return savedOrder;
@@ -54,7 +64,16 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     public Order acceptOrder(Order order) {
-        //TODO 30.05.2020 validate order
+        if (!order.getProducts().stream()
+                .flatMap(product -> product.getUsedIngredients().stream())
+                .allMatch(ingredientService::checkForAvailability)) {
+            throw new RuntimeException("We don't have some ingredients for this order. Request some from storage!");
+        }
+
+        order.getProducts().stream()
+                .flatMap(product -> product.getUsedIngredients().stream())
+                .forEach(ingredientService::takeIngredientsFromBarStand);
+
         orderWebSocketMessageSender.sendAcceptedOrder(order.setStatus(Status.ACCEPTED));
         deleteOrderById(order.getId());
         return order;
